+++ src/Refitter.Tests/OpenApi/OpenApiDocumentFactoryMergeTests.cs
@@ -5,9 +5,8 @@ using NJsonSchema;
 using NSwag;
 using Refitter.Core;
 using Refitter.Tests.Resources;
-using TUnit.Core;
 
-namespace Refitter.Tests;
+namespace Refitter.Tests.OpenApi;
 
 public class OpenApiDocumentFactoryMergeTests
 {
